6-size.c

+++ hello_world/6-size.c
@@ -11,7 +11,7 @@ int main(void)
 printf("Size of char: %lu bytes\n", sizeof(char));
 printf("Size of int: %lu bytes\n", sizeof(int));
 printf("Size of long int: %lu bytes\n", sizeof(long int));
-printf("Size of long long int : %lu bytes\n", sizeof(long long int));
+printf("Size of long long int : %ld bytes\n", sizeof(unsigned long int));
 printf("Size of float: %lu bytes\n", sizeof(float));
 return (0);
 }
